Reduce lock overhead for empty shards Add task callback class to Message

+++ src/main/java/com/netflix/astyanax/recipes/queue/MessageQueueEntry.java
@@ -90,7 +90,7 @@ public class MessageQueueEntry {
         return MessageQueueEntryState.values()[state];
     }
 
-    public short getPriority() {
+    public byte getPriority() {
         return priority;
     }
